Rewrite pipeline using sklearn
* Add reformat.R to convert smartsurvey format.
* Add feature generators.
* Extract data from db not flat files.
* Update README.
* Add db_to_flat_file.py to create merged db dump.
* Add urllookup as module.
* Add urllookup tests on travis.
* Add PII removal using scrubadub.
* Dump to pickle objects.

Note: DataFrameSelector outputs a numpy array, but it is easier to calculate
date features using pandas prior to conversion to a numpy array. For
this reason, create an argument as to whether the DataFrameSelector is
dealing with date features or not, and then apply the date transforms if
so, and output a numpy array.

Some weirdness around installing scikit-learn.
See:

* https://github.com/scikit-learn/scikit-learn/issues/8242
* https://github.com/scikit-learn/scikit-learn/issues/7867

Solution is to force numpy and scipy installation before
requirements.txt is installed.